% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date.R
\name{convert_data}
\alias{convert_data}
\title{Helper Function that allows the conversion of date variables, often find in
quote tables from data providers. i.e.when  'quote_datetime' the form of
'YYYY-MM-DD HH:MM:SS'}
\usage{
convert_data(Data, dmY = FALSE)
}
\arguments{
\item{data}{Dataframe incl. variables 'expiration' and 'quote_datetime'.
'dmY==TRUE' should be used if data is int the format of d-m-Y.}
}
\value{
Dataframe with new variables 'QuoteTime' (indicating minutes from
  midnight until quote) and 'QuoteDate'
}
\description{
Helper Function that allows the conversion of date variables, often find in
quote tables from data providers. i.e.when  'quote_datetime' the form of
'YYYY-MM-DD HH:MM:SS'
}
\examples{
convert_data(data,dmY=TRUE) 
}
\seealso{
\code{\link{letVixShiny}},\code{\link{vic_cal}}
}
