% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letVixShiny.R
\name{letVixShiny}
\alias{letVixShiny}
\title{Create a shiny app for the VIX data}
\usage{
letVixShiny(vix)
}
\arguments{
\item{vix}{An object of data type VIX, containing vix qotes (i.e. output of vic_calc)}
}
\value{
Siny App
}
\description{
Create a shiny app for the VIX data
}
\examples{
letVixShiny(vix)
}
\seealso{
\code{\link{vic_calc}}, \code{\link{convert_date}}
}
