% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvix.R
\name{plotvix}
\alias{plotvix}
\title{Plot vix_calc output}
\usage{
plotvix(vix, smoother = FALSE, type = "l")
}
\arguments{
\item{vix}{Output from \code{\link{vix_calc}}}

\item{smoother}{eg. "lm", "glm", "gam", "loess". FALSE for no smoother (default)}

\item{type}{"l" for line, "p" for points}
}
\value{
Nothing: just a nice plot
}
\description{
Plot vix_calc output
}
\examples{
vix <- data.frame('VIX' = runif(450, 9, 9.7))
plotvix(vix, smoother="loess",type="p")
}
\seealso{
\code{\link{vix_calc}, \link{fitVIX}, \link{letVixShiny}}
}
