% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vixr.R
\docType{package}
\name{vixr}
\alias{vixr}
\alias{vixr-package}
\title{vixr: A package for replicating the VIX}
\description{
The vixr package contains four categories of important functions. For more details, please consult the vignette.
}
\seealso{
\code{\link{vix_calc}, \link{fitVIX},\link{plotvix},\link{letVixShiny}}
}
