% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVIX.R
\name{fitVIX}
\alias{fitVIX}
\title{fitVIX - Fit a basic statistical model to VIX data}
\usage{
fitVIX(Data, fit_type = c("lm", "loess", "smooth.spline"))
}
\arguments{
\item{Data}{A VIX object (returned by vix_calc())}

\item{fit_type}{The type of model required, either linear regression (\code{lm}), loess (\code{loess}), or smoothing spline (\code{smooth.spline})}
}
\value{
Returns a list of class fitVIX which includes the model details as well as fit type used
}
\description{
fitVIX - Fit a basic statistical model to VIX data
}
\examples{
SampleData <- data.frame('VIX' = runif(450, 9, 9.7))
fitVIX(SampleData, fit_type = 'smooth.spline')

}
\seealso{
\code{\link{letVixShiny}, \link{vix_calc}}
}
