% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letVixShiny.R
\name{letVixShiny}
\alias{letVixShiny}
\title{Create a shiny app for the VIX data}
\usage{
letVixShiny(vix)
}
\arguments{
\item{vix}{An object of data type VIX, containing vix qotes (i.e. output of vic_calc)}
}
\value{
Siny App
}
\description{
Create a shiny app for the VIX data
}
\examples{
\dontrun{
vix <- data.frame('VIX' = runif(450, 9, 9.7))
letVixShiny(vix)
 }
}
\seealso{
\code{\link{vix_calc}, \link{fitVIX}}
}
