% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vix_calc.R
\name{vix_calc}
\alias{vix_calc}
\title{Vix_Calc: Calculating volatility index quotes}
\usage{
vix_calc(Data, rf_near = 0.002, rf_next = 0.002, nearT = 28, nextT = 35)
}
\arguments{
\item{Data}{Dataframe incl. 'QuoteDate' (format:YYYY-MM-DD), 'QuoteTime' (format:hh:mm:ss), 'expiration' (format:YYYY-MM-DD), 'option_type' ('C' for call or 
'P' for put), 'bid', 'ask','active_underlaying_price' (S&P 500 spot prices),'underlying symbol' with '^SPX' representing the S&P500 symbol}

\item{rf_near}{Defining risk free rate for near term}

\item{rf_next}{Defining risk free rate for next term}

\item{nearT}{Defining the near term}

\item{nextT}{Defining the next term}
}
\value{
A vector of VIX quptes
}
\description{
Vix_Calc: Calculating volatility index quotes
}
\examples{
vix_calc(VIX_SampleData, rf_near = 0.002,rf_next = 0.002, nearT=28, nextT=35) 

}
\seealso{
\code{\link{letVixShiny}, \link{fitVIX},\link{plotvix}}
}
